+++ inst/app-pixfigure/app.R
@@ -9,14 +9,35 @@ ui <- fluidPage(
     sidebarPanel(
       width = 3,
       textAreaInput("path", "Image paths", value = path, height = "200px"),
-      textInput("sep", "Separator", value = ";"),
-      checkboxInput("caption_check", "Use captions", value = FALSE),
+      textInput("sep", "Path and caption separator", value = ";"),
+      checkboxInput("caption_check", "Use caption", value = FALSE),
       uiOutput("caption_ui"),
       checkboxInput("link", "Use lightbox", value = TRUE),
+      checkboxInput("caption_lightbox", "Use lightbox caption", value = TRUE),
       textInput("h", label = "Height", value = "auto"),
       textInput("w", label = "Width", value = "100%"),
-      selectInput("fit", label = "Object fit", choices = c("cover", "contain", "fill", "none", "scale-down"), selected = "cover"),
-      selectInput("position", label = "Object position", choices = c("center", "top", "bottom", "left", "right", "top left", "top right", "bottom left", "bottom right"), selected = "center"),
+      selectInput(
+        "fit",
+        label = "Object fit",
+        choices = c("cover", "contain", "fill", "none", "scale-down"),
+        selected = "cover"
+      ),
+      selectInput(
+        "position",
+        label = "Object position",
+        choices = c(
+          "center",
+          "top",
+          "bottom",
+          "left",
+          "right",
+          "top left",
+          "top right",
+          "bottom left",
+          "bottom right"
+        ),
+        selected = "center"
+      ),
       HTML("Image from <a href='https://www.pexels.com/'>Pexels.</a>")
     ),
     mainPanel(
@@ -40,7 +61,7 @@ server <- function(input, output) {
 
     if (input$caption_check) {
       if (!is.null(input$caption)) {
-        cpt <- unlist(strsplit(input$caption, ";"))
+        cpt <- unlist(strsplit(input$caption, input$sep))
         if (length(cpt) != length(path)) {
           stop("Number of captions do not match number of images.")
         }
@@ -48,6 +69,7 @@ server <- function(input, output) {
           path,
           caption = cpt,
           link = input$link,
+          caption_lightbox = input$caption_lightbox,
           h = input$h,
           w = input$w,
           fit = input$fit,
@@ -58,6 +80,7 @@ server <- function(input, output) {
       pixture::pixfigure(
         path,
         link = input$link,
+        caption_lightbox = input$caption_lightbox,
         h = input$h,
         w = input$w,
         fit = input$fit,